Create ULipSyncGameObject.cs
It's sort of like ULipSyncTexture but for game objects. 
I use this for 2D sprite puppets, like those imported via the PSB importer will have different origins for their phoneme sprites due to the automatic sprite slicing. Not only that, but you'll already have a sprite hierarchy you want to keep, so this requires no edits to the layer structure in your PSB or PSD imports. Very convenient.